+++ wok_code/wok_code/egg-info/CHANGELOG.rst
@@ -3,6 +3,8 @@
 
 * [FIX] odoo_translation.py: case correction
 * [FIX] run_odoo_debug: sometimes crashes on OCB/addons modules
+* [FIX] gen_readme.py: Odoo repository documentation
++ [FIX[ please docs: count assertions
 
 2.0.15 (2024-02-17)
 ~~~~~~~~~~~~~~~~~~~